refactoring test code

+++ investment-service/src/main/java/project/investmentservice/api/InvestmentApiController.java
@@ -45,7 +45,7 @@ public class InvestmentApiController {
             
             double averagePrice = usersStock.getAveragePrice();
             Long quantity = usersStock.getQuantity();
-            Long seedMoney = user.getSeedMoney();
+            double seedMoney = user.getSeedMoney();
 
             return new PurchaseStockResponse(SUCCESS, averagePrice, quantity, seedMoney);
         }
